rename: Modal 컴포넌트 경로 이동 (common 디렉터리 내부로 이동)
