+++ components/src/lib/components/common/Modal/ModalContent.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { StyleDirection } from '../types';
+import { StyleDirection } from '../../types';
 
 export interface ModalContentProps {
   message?: string;
