+++ components/src/lib/components/common/Modal/ModalMain.tsx
@@ -1,12 +1,12 @@
 import React, { MouseEvent } from 'react';
 
 import styled from 'styled-components';
-import { MODAL_SIZE } from '../constants/styles';
-import { StyleSize, StylePosition } from '../types';
+import { MODAL_SIZE } from '../../constants/styles';
+import { StyleSize, StylePosition } from '../../types';
 
-interface ModalProps {
+export interface ModalProps {
   $size?: StyleSize;
-  $position: StylePosition;
+  $position?: StylePosition;
   children?: React.ReactNode;
   onCloseModal: () => void;
 }
