+++ components/src/lib/components/common/Modal/index.tsx
@@ -2,8 +2,8 @@ import ModalMain from './ModalMain';
 import ModalHeader from './ModalHeader';
 import ModalContent from './ModalContent';
 import ModalFooter from './ModalFooter';
-import Button from '../common/Button';
-import Input from '../common/Input';
+import Button from '../Button';
+import Input from '../Input';
 
 const Modal = Object.assign(ModalMain, {
   Header: ModalHeader,
