+++ components/src/lib/index.ts
@@ -1,2 +1,2 @@
-export { default as Modal } from './components/Modal';
+export { default as Modal } from './components/common/Modal';
 export { default as useModal } from './hooks/useModal';
